as uint8

+++ heavyhex_qft/utils.py
@@ -5,8 +5,8 @@ from qiskit.transpiler import CouplingMap
 
 def as_bitarray(bitstr: str | np.ndarray):
     if isinstance(bitstr, str):
-        return np.array(list(map(int, bitstr[::-1])))
-    return np.asarray(bitstr)
+        return np.array(list(map(int, bitstr[::-1])), dtype=np.uint8)
+    return np.asarray(bitstr, dtype=np.uint8)
 
 
 def to_pauli_string(link_ops: dict[int, str], num_qubits: int) -> str:
